feat: 优先级1修复完成 - CopyToFile改用uint64_t支持>4GB数据
【第八轮修复】完成优先级1的关键改造：

1. CopyToFile方法参数改为uint64_t
   - Common/ReceiveCacheService.h: 参数声明改为uint64_t&
   - Common/ReceiveCacheService.cpp: totalCopied/targetSize改为uint64_t
   - 支持处理>4GB的文件接收和保存

2. SaveReceiveDataToFile同步更新
   - 参数bytesWritten改为uint64_t
   - CString::Format占位符改为%llu（支持uint64_t）

3. 编译验证
   - ✅ 编译成功：0 error / 0 warning
   - 耗时：17.51秒

【优先级2-5】制定详细的修订方案文档
   - 优先级2：2GB限制与流式发送策略（两个选项）
   - 优先级3：运行库配置（/MT /MTd）
   - 优先级4：资源清理（PortMaster_Icon.ico）
   - 优先级5：端口占用检测（IsPortAvailable实现）

对应验证报告：PortMaster_修订校验报告_20251022-143638.md
对应分析报告：第13.3、13.4、13.6、13.8节

+++ Common/ReceiveCacheService.h
@@ -149,7 +149,7 @@ public:
 	 * - 若临时文件不存在或未初始化，返回false并记录日志
 	 * - 成功后通过bytesWritten参数返回实际写入的字节数
 	 */
-	bool CopyToFile(const std::wstring& targetPath, size_t& bytesWritten);
+	bool CopyToFile(const std::wstring& targetPath, uint64_t& bytesWritten);
 
 	/**
 	 * @brief 获取内存缓存数据（已删除）
